refactor: split dashboard page into focused hooks and components
